Split rewards into quest and vendor, cleaned up IDs, added docs
